Fix: Check availability with workers restored
